Add a backward-compatible recursive stacked navigation bar.
From any page on a site, every level of the directory hierarchy is
shown, toplevel first, each level exactly as before but without the
now-redundant "." and ".." links but instead (CSS-overridable)
highlighting of every component of the current page's path.
Additionally, any link to the current page is now a differently-colored
non-link.

This behavior is enabled by the rawkrc option "recursive_menu" (name
chosen arbitrarily; please advise of better alternatives).